Refactor: Separate demo script from expert_system.py library
Moved the `demo_expert_system` function and the `if __name__ == "__main__":` block from `src/blank_business_builder/expert_system.py` to a new standalone script `simple_expert_demo.py` in the root directory. This improves code modularity and prevents library code from running as a script.

Also removed an unused `json` import from `expert_system.py`.

Verified by running the new demo script and existing tests.
